Create README.md
3-D modelling is the use of software to create a virtual three-dimensional model of some physical object. These objects can be made using several tools like polygon primitives, spline curves, NURBS(non-uniform Rational b-spline).
The main aim of building up these characters was to make a short film out of it.
This short film consists of a small child, his mother, a beggar, and a priest. 
The main message which I tried to pass through this film is discrimination between Hindu and Muslim cast. As we can see this is one of the main issues in  India. 
I have created these 4 models using Autodesk Maya software as it was a new and advanced software I faced a lot of struggle in every step I towards building the character my main aim is to build the desired short movie after I learn character animation.

Models:
As I have written earlier there are 4 characters in this movie and this is how they look like shown below:
The steps that were performed for building this character are :
1.	First, a reference image was taken so that we can build equal a proper model
2.	After that, a polygon primitive was taken and adjusted to the desired size
3.	After that different parts of the body like the head ears etc. were combined and then connected using the merge tool.
4.	And then the texturing part took place.

Conclusion:
After these characters are rendered they can further be used for animation for making movie clips, or even games etc.